Testing done Role, user, warehouse, outlet

+++ src/main/java/com/digitalrangersbd/DeepManage/Service/RoleService.java
@@ -1,8 +1,10 @@
 package com.digitalrangersbd.DeepManage.Service;
 
+import com.digitalrangersbd.DeepManage.Authorization.RoleAuthorization;
 import com.digitalrangersbd.DeepManage.Dto.RoleDto;
 import com.digitalrangersbd.DeepManage.Dto.RoleUpdateDto;
 import com.digitalrangersbd.DeepManage.Entity.Role;
+import com.digitalrangersbd.DeepManage.JWT.UserContext;
 import com.digitalrangersbd.DeepManage.Repository.RoleRepository;
 import org.springframework.stereotype.Service;
 import java.time.LocalDate;
@@ -14,18 +16,25 @@ import java.util.Optional;
 public class RoleService {
 
     private final RoleRepository roleRepository;
+    private final RoleAuthorization roleAuthorization;
+    public RoleService(RoleRepository roleRepository, RoleAuthorization roleAuthorization) {
 
-    public RoleService(RoleRepository roleRepository) {
         this.roleRepository = roleRepository;
+        this.roleAuthorization = roleAuthorization;
     }
 
     //Create new role
     public Role createRole(RoleDto dto) {
 
+        String userId = UserContext.getUserId();
+
+        if(roleAuthorization.hasCreateRolePermission(userId)){
+            throw new SecurityException("User does not have the permission to create role.");
+        }
+
         Role role = new Role();
         role.setName(dto.getName());
         role.setCreated_by_id(dto.getCreated_by_id());
-        role.setCreated_by_name(dto.getCreated_by_name());
         role.setPermission(dto.getPermission());
 
         // Set timestamps
@@ -40,23 +49,43 @@ public class RoleService {
 
     //Get roles
     public List<Role> getAllRoles() {
+
+        String userId = UserContext.getUserId();
+        System.out.println("this is the user id:"+userId);
+        if(roleAuthorization.hasViewRolePermission(userId)){
+            throw new SecurityException("User does not have the permission to view role.");
+        }
+
         return roleRepository.findAll();
     }
 
     //Get roles by id
     public Optional<Role> getRolesById(String id) {
+
+        String userId = UserContext.getUserId();
+        if(roleAuthorization.hasViewRolePermission(userId)){
+            throw new SecurityException("User does not have the permission to view role.");
+        }
+
         return roleRepository.findById(id);
     }
 
     //update role
     public Role updateRole(String id, RoleUpdateDto dto) {
 
+        String userId = UserContext.getUserId();
+        if(roleAuthorization.hasUpdateRolePermission(userId)){
+            throw new SecurityException("User does not have the permission to update role.");
+        }
+
         return roleRepository.findById(id)
                 .map(existingRole -> {
                     if (dto.getName() != null) existingRole.setName(dto.getName());
                     if (dto.getCreated_by_id() != null) existingRole.setCreated_by_id(dto.getCreated_by_id());
-                    if (dto.getCreated_by_name() != null) existingRole.setCreated_by_name(dto.getCreated_by_name());
-                    if (dto.getPermission() != null) existingRole.setPermission(dto.getPermission());
+                    if (dto.getPermission() != null) {
+                        existingRole.getPermission().clear();
+                        existingRole.setPermission(dto.getPermission());
+                    }
 
                     existingRole.setUpdated_date(LocalDate.now());
                     existingRole.setUpdated_time(LocalTime.now());
@@ -68,6 +97,12 @@ public class RoleService {
 
     //delete role
     public boolean deleteRole(String id) {
+
+        String userId = UserContext.getUserId();
+        if(roleAuthorization.hasDeleteRolePermission(userId)){
+            throw new SecurityException("User does not have the permission to delete role.");
+        }
+
         if (roleRepository.existsById(id)) {
             roleRepository.deleteById(id);
             return true;
